Move verify() to SignatureBundle.verify()
